QoL changes for integration reruns (#45)
* add logging statements

* use zarr partial reading for assertions

* use original barcode plot implementation per study

* fix random seed for subsetting

* update relabel script to use backed mode for h5ad or partial reading for zarr, output to linked zarr

* further optimise preprocessing file linking

* update umap outlier parameter

* only dataset-specific line in datasets.tsv as parameter

* solve passing args for no HVG vs default HVG

* more efficient h5ad vs zarr file handling

* update to latest AIDA dataset and Powell annotations

* force annotation index to string

* remove index from dataset df slice, to make independent from order in dataset.tsv

* rename profile to czbiohub

* bugfix conserver adata dimensions during HVG

* use same dataset getter function for modules

* more efficient testing

* add logging statements

* use zarr file for subsetting

* use only first column in mapping hierarchy to map all labels

* cleaner config for exploration preprocessing

* use preprocessing pipeline for per lineage integration

* update plot to include numbers of cells

* adapt to subsetting data to HVGs

* use rapids_singlecell for clustering

* update scanpy_rapids env

* ignore software env updates

* add missing resource

* save raw counts in adata.raw

* smaller batch size

* change output dir folder structure

* link zarr files to relative paths by default

* refactor datasets per module getter

* rename CL label

* fix relative path linking (from absolute output path)

* use read_elem

* bugfix: select layer condition inverted

* remove per batch filter

* rearrange operations for splitting per lineage

* default to outer join for genes

* add logging statements

+++ workflow/utils/io.py
@@ -1,3 +1,4 @@
+import os
 from pathlib import Path
 import shutil
 import anndata as ad
@@ -25,7 +26,7 @@ def read_anndata_or_mudata(file):
         return read_anndata(file)
 
 
-def link_zarr(in_dir, out_dir, file_names=None, overwrite=False):
+def link_zarr(in_dir, out_dir, file_names=None, overwrite=False, relative_path=True):
     """
     Link to existing zarr file
     """
@@ -44,9 +45,18 @@ def link_zarr(in_dir, out_dir, file_names=None, overwrite=False):
             continue
         new_file = out_dir / f.name
         if overwrite and new_file.exists():
-            print(f'Directory {new_file} exists, overwrite...')
-            if new_file.is_dir():
+            print(f'Replace {new_file} with link')
+            if new_file.is_dir() and not new_file.is_symlink():
                 shutil.rmtree(new_file)
             else:
                 new_file.unlink()
-        new_file.symlink_to(f.resolve())
+        
+        path_to_link_to = f.resolve()
+        if relative_path:
+            path_to_link_to = Path(
+                os.path.relpath(
+                    path_to_link_to,
+                    new_file.parent.resolve()
+                )
+            )
+        new_file.symlink_to(path_to_link_to)
